+++ test.sh
@@ -22,6 +22,8 @@ swiftc -parse-as-library "$@" \
     Sources/AIUsage/WorkSchedule.swift \
     Sources/AIUsage/Pace.swift \
     Sources/AIUsage/ClaudeProfile.swift \
+    Sources/AIUsage/CodexProfile.swift \
+    Sources/AIUsage/CodexAccountAccess.swift \
     Sources/AIUsage/Fetcher.swift \
     Sources/AIUsage/ProviderFolderAccess.swift \
     Sources/AIUsage/BrandGlyph.swift \